Added Timer
included timer and improved text display, gonna have a menu come up after timer hits zero

+++ SSC-RowdyHacks2022/Assets/Scripts/FPSCounter.cs
@@ -17,6 +17,6 @@ public class FPSCounter : MonoBehaviour
     void Update()
     {
         float fps = 1 / Time.unscaledDeltaTime;
-        framesDisplay.text = "" + fps;
+        framesDisplay.text = "" + (int)fps;
     }
 }
